Add app icon
The icon is a fast-forward symbol on a red button, drawn with
perspective. The icon follows the GNOME app icon guidelines and uses the
GNOME color palette. The icon is inspired by the YouTube and NewPipe
icons. The two arrows of the fast-forward symbol resemble the two Ds of
the DewDuct name.